+++ lib/rearmed.rb
@@ -5,10 +5,42 @@ require 'rearmed/exceptions'
 
 module Rearmed
 
-  @enabled_patches = eval(File.read(File.join(File.dirname(__FILE__), 'rearmed/default_enabled_patches.hash')))
+  DEFAULT_PATCHES = {
+    array: {}, 
+    date: {}, 
+    enumerable: {}, 
+    hash: {}, 
+    integer: {}, 
+    object: {}, 
+    string: {}, 
+  }.freeze
+  private_constant :DEFAULT_PATCHES
+  
+  @enabled_patches = Marshal.load(Marshal.dump(DEFAULT_PATCHES)) 
 
   def self.enabled_patches=(val)
-    @enabled_patches = val
+    if val.nil?
+      @enabled_patches = Marshal.load(Marshal.dump(DEFAULT_PATCHES)) 
+    elsif val == :all
+      @enabled_patches = val
+    elsif val.is_a?(Hash)
+      keys = DEFAULT_PATCHES.keys
+
+      keys.each do |k|
+        methods = val[k] || val[k.to_sym]
+        if methods
+          if methods.is_a?(Hash) || methods == true
+            @enabled_patches[k] = methods
+          else
+            raise TypeError.new('Invalid value within the hash passed to Rearmed.enabled_patches=')
+          end
+        else
+          @enabled_patches[k] = {}
+        end
+      end
+    else
+      raise TypeError.new('Invalid value passed to Rearmed.enabled_patches=')
+    end
   end
 
   def self.enabled_patches