+++ lib/rearmed/monkey_patches/date.rb
@@ -1,4 +1,4 @@
-if Rearmed.enabled_patches[:date] == true || Rearmed.dig(Rearmed.enabled_patches, :date, :now)
+if Rearmed.enabled_patches == :all || Rearmed.enabled_patches[:date] == true || Rearmed.dig(Rearmed.enabled_patches, :date, :now)
 
   require 'date'